Py/requirements.txt
openai-whisper
pyaudio
keyboard
pyautogui
pynput
SpeechRecognition
python-Levenshtein
duckdb